youtube-data-processing/requirements.txt
psycopg2-binary
celery
redis
uvicorn